Patch OSL for connection to color components.
The support from OSL commit 2398438c0c15f48d2aba2148c5c18286b8b371f1 is
buggy when not all components are connected and the sources are
different layers.

The reason is that the initialization of the partially connected input
is done from the source layer when it detects that not all target
components are connected to. This is excessively complex and buggy when
there is more than one source layer as the second layer to run
overwrites the output of the first layer (in the input of the target
layer).

Now, we instead track which inputs have full connection (not per
component) to them. Inputs which don't have a full connection get their
initialization run as usual, in the target layer, which allows it to run
only once and before any of the upstream layers write overriding values
because of the component connections. So it all just works.

The only downside is that the case where all the components are
connected separately is no longer tracked. The old code was detecting
this and skipping the initialization code. While it's a nice thing to do
in theory, I suspect it is of little practical value, at least for the
way we use OSL here (our initializers never fetch user values).