Patch OSL to support connecting to array elements.
Appears to work reasonably well, at least for our intended use case.

The one major caveat is that when connecting to an unsized array (eg.
float x[]), its size must be set by a parameter for connections to be
done to elements. Even if all the elements of the parameters will be
overwritten by connections.

This is because setting the size causes storage for it to be allocated
in the shader instance and it would be difficult to do that
incrementally as more connections to ever higher indices show up.
There might also be problems initializing elements without connections.